Added unit test and fixed reversed results
