+++ package.json
@@ -5,7 +5,7 @@
   "description": "Compare the performance of Rust project branches using criterion",
   "main": "main.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest",
     "build": "ncc build main.js --minify"
   },
   "keywords": [],
@@ -18,6 +18,7 @@
   },
   "devDependencies": {
     "@vercel/ncc": "^0.36.0",
+    "jest": "^29.5.0",
     "prettier": "2.8.3"
   }
 }
